+++ src/read_line.c
@@ -1,20 +1,26 @@
 #include "shell.h"
 
-char *read_line(void) {
-    char *line = NULL;
-    size_t bufsize = 0;
+char *read_line(void)
+{
+    char *line = NULL; // Pointer to store the line read from stdin
+    int bufsize = 0;   
 
-    if (getline(&line, &bufsize, stdin) == -1) {
-        if (feof(stdin)) {
-            free(line);
+    // Read a line from stdin
+    if (getline(&line, &bufsize, stdin) == -1)
+    {
+        // Check if getline failed due to end-of-file
+        if (feof(stdin))
+        {
+            free(line);        
             exit(EXIT_SUCCESS);
-        } else {
+        }
+        else
+        {
+            // Print error message for getline failure
             perror("getline failed");
-            free(line);
-            exit(EXIT_FAILURE);
+            free(line);         
+            exit(EXIT_FAILURE); 
         }
     }
-    return line;
+    return line; // Return the line read from stdin
 }
-
-
